requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
langgraph==0.2.0
langsmith==0.1.75
kafka-python==2.0.2
qdrant-client==1.9.0
asyncpg==0.29.0
httpx==0.27.0
pydantic==2.7.0
pytest==8.2.0
pytest-asyncio==0.23.0
python-dotenv==1.0.0
